requirements.txt
Fabric==1.8.1
Pillow==2.5.3
ecdsa==0.10
gevent==0.13.8
greenlet==0.4.1
httplib2==0.8
oauth2==1.5.211
oauthlib==0.6.0
paramiko==1.12.1
pycrypto==2.6.1
python-twitter==1.1
requests==2.0.0
requests-oauthlib==0.4.0
simplejson==3.3.1
wsgiref==0.1.2
